Move the only StoreTest at the top level  more targeted StoreTest
