+++ tests/phpunit/includes/storage/StoreTest.php
@@ -198,4 +198,24 @@ class StoreTest extends DatabaseTestCase {
 		);
 	}
 
+	public function testGetRedirectTarget() {
+
+		$wikipage = new DIWikiPage( 'Foo', NS_MAIN );
+		$expected = new DIWikiPage( 'Bar', NS_MAIN );
+
+		$instance = $this->getMockBuilder( '\SMW\Store' )
+			->disableOriginalConstructor()
+			->setMethods( [ 'getPropertyValues' ] )
+			->getMockForAbstractClass();
+
+		$instance->expects( $this->once() )
+			->method( 'getPropertyValues' )
+			->will( $this->returnValue( [ $expected ] ) );
+
+		$this->assertEquals(
+			$expected,
+			$instance->getRedirectTarget( $wikipage )
+		);
+	}
+
 }
